# api-spec.yaml
openapi: 3.0.3
info:
  title: Minecraft Server Manager API
  description: RESTful API for managing Minecraft servers, user authentication, and administrative functions
  version: 0.1.0-alpha
  contact:
    name: Minecraft Server Manager
    url: https://github.com/your-username/mcServerManager
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://your-domain.com/api/v1
    description: Production server
security:
  - sessionAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Login user and create session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  example: securepassword
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get current user information
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password
      description: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_password
                - new_password
                - confirm_password
              properties:
                current_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password
                confirm_password:
                  type: string
                  format: password
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/setup:
    post:
      tags:
        - Authentication
      summary: Setup admin account
      description: Create initial admin account (first run only)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - confirm_password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                confirm_password:
                  type: string
                  format: password
                email:
                  type: string
                  format: email
                  example: admin@example.com
      responses:
        '201':
          description: Admin account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset user password (admin or self)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - new_password
                - confirm_password
              properties:
                user_id:
                  type: integer
                  description: User ID (admin only)
                new_password:
                  type: string
                  format: password
                confirm_password:
                  type: string
                  format: password
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Server Management API
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error description
        error:
          type: string
          example: Detailed error information

    # User schemas
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@example.com
        is_admin:
          type: boolean
          example: true
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2025-09-04T16:20:00.000Z"
        last_login:
          type: string
          format: date-time
          example: "2025-09-04T16:20:00.000Z"
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
...
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
...
    SetupResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                username:
                  type: string
                  example: admin
                is_admin:
                  type: boolean
                  example: true
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...